call api key environment value
