+++ Sources/ChatXPN/Chat/Chat.swift
@@ -10,10 +10,9 @@ import NotraAuth
 
 struct Chat: View {
     @StateObject private var chatVM = ChatViewModel()
-    let callApiKey: String
     
     var body: some View {
-        ChatList(callApiKey: callApiKey)
+        ChatList()
             .navigationTitle("")
             .navigationBarTitleDisplayMode(.inline)
             .environmentObject(chatVM)
@@ -33,5 +32,5 @@ struct Chat: View {
 }
 
 #Preview {
-    Chat(callApiKey: "")
+    Chat()
 }
