+++ Sources/ChatXPN/Chat/ChatListCell.swift
@@ -11,11 +11,10 @@ import NotraAuth
 struct ChatListCell: View {
     @EnvironmentObject var viewRouter: ChatViewRouter
     let chat: ChatModelViewModel
-    let callApiKey: String
     
     var body: some View {
         Button {
-            viewRouter.pushChatPath(.chatRoom(chat: chat, apiKey: callApiKey))
+            viewRouter.pushChatPath(.chatRoom(chat: chat))
         } label: {
             HStack(alignment: .top, spacing: 14) {
                 
@@ -54,6 +53,6 @@ struct ChatListCell: View {
 }
 
 #Preview {
-    ChatListCell(chat: PreviewModels.chats[0], callApiKey: "")
+    ChatListCell(chat: PreviewModels.chats[0])
         .environmentObject(ChatViewRouter())
 }
